Add cron job for DB sync at KPNO.
This adds a high-level script that is very short and simply loads
the latest stable desiconda / desimodules.  This is kept in the
etc directory and manually copied to ~datasystems/desimodel_sync
at KPNO.  The other script (`desimodel_sync_kpno.sh`) is installed
with desimodel.  This finds the latest calibration file and runs
the python sync script.  Tested at KPNO with the --test option
and verified that the log files are created correctly and that
the underlying sync command runs as expected.