Update footprint.py
Added the `workers` optional argument to the relevant functions and set default to -1. Also wrote `distance_upper_bound = threshold` in `is_point_in_desi` as this allows to prune the tree further. As a result the points in desi will return an infinite distance, which invalidates the previous check `d < threshold` so I changed it to `np.isfinite(d)`.